+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added the quantity to the accounts tab of the holding detail dialog (experimental)
+
 ### Fixed
 
 - Fixed the allocation in the accounts tab of the holding detail dialog caused by floating-point rounding
+- Fixed the allocation in the accounts tab of the holding detail dialog by excluding the cash balance of the account
 - Fixed the account aggregations in impersonation mode to be based on the impersonated user
 - Fixed the base currency of the activities in impersonation mode to be based on the impersonated user
 - Fixed the base currency of the dividends in impersonation mode to be based on the impersonated user